cambios en las particulas
 - de dispercion radio, ahora es radio_dispercion.
 - se puede seleccionar que las particulas tengan color random.
 - La particula ahora desaparece cuando exede la distancia maxima desde donde aparecio, y no desde donde esta el spawn_pos.

+++ particle.py
@@ -57,7 +57,7 @@ class Particle:
         return self.__radio
     @radio.setter
     def radio(self,radio):
-        self.__radio = int(radio)
+        self.__radio = float(radio)
         self.generate()
 
     @property
